feat(env-check): pr comments
