+++ packages/environment-check/src/index.ts
@@ -1,4 +1,3 @@
 export * from './types';
 export * from './checks';
-export { convertResultsToMarkdown, storeResultsZip, archiveProject } from './output';
-export { formatResultsForClipboard } from './formatter';
+export { convertResultsToMarkdown, storeResultsZip, archiveProject, convertResultsToString } from './output';
