+++ packages/environment-check/src/output/index.ts
@@ -1,2 +1,3 @@
 export * from './markdown';
 export * from './zip';
+export * from './string';
